basic support for go1.15
