+++ objc/objc_runtime_darwin_test.go
@@ -5,6 +5,7 @@ package objc_test
 
 import (
 	"fmt"
+	"log"
 	"unsafe"
 
 	"github.com/ebitengine/purego"
@@ -14,12 +15,14 @@ import (
 func ExampleAllocateClassPair() {
 	var class = objc.AllocateClassPair(objc.GetClass("NSObject"), "FooObject", 0)
 	class.AddMethod(objc.RegisterName("run"), objc.NewIMP(func(self objc.ID, _cmd objc.SEL) {
+		log.Println(self)
 		fmt.Println("Hello World!")
 	}), "v@:")
 	class.Register()
 
 	var fooObject = objc.ID(class).Send(objc.RegisterName("new"))
 	fooObject.Send(objc.RegisterName("run"))
+	log.Println(fooObject)
 	// Output: Hello World!
 }
 
@@ -43,13 +46,13 @@ func ExampleClass_AddIvar() {
 }
 
 func ExampleIMP() {
-	imp := objc.NewIMP(func(self objc.ID, _cmd objc.SEL) {
-		fmt.Println("IMP:", self, _cmd)
+	imp := objc.NewIMP(func(self objc.ID, _cmd objc.SEL, a3, a4, a5, a6, a7, a8, a9 int) {
+		fmt.Println("IMP:", self, _cmd, a3, a4, a5, a6, a7, a8, a9)
 	})
 
-	purego.SyscallN(uintptr(imp), 105, 567)
+	purego.SyscallN(uintptr(imp), 105, 567, 9, 2, 3, ^uintptr(4), 4, 8, 9)
 
-	// Output: IMP: 105 567
+	// Output: IMP: 105 567 9 2 3 -5 4 8 9
 }
 
 func ExampleID_SendSuper() {
